backend/requirements.txt
# backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.35
supabase==2.4.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
requests==2.32.3
Pillow==10.4.0
opencv-python-headless==4.10.0.84
psycopg[binary]==3.2.1